+++ src/controllers/tonos-cli/components.ts
@@ -1,0 +1,22 @@
+import * as path from "path";
+import {
+    Component,
+    tondevHome,
+} from "../../core";
+
+function tonoscliHome() {
+    return path.resolve(tondevHome(), "tonos-cli");
+}
+
+export const components = {
+    tonoscli: new class extends Component {
+        getSourceName(version: string): string {
+            return `tonos-cli-${version.split(".").join("_")}-{p}.zip`;
+        }
+    }(tonoscliHome(), "tonos-cli", {
+        extractCurrentVersionRegExp: /tonos_cli\s*([0-9.]+)/,
+        executable: true,
+        globally: true,
+    }),
+};
+