+++ src/core/utils.ts
@@ -1,11 +1,15 @@
 import * as path from "path";
 import * as os from "os";
 import * as fs from "fs";
-import { spawn, SpawnOptionsWithoutStdio } from "child_process";
-import * as http from "http";
+import {
+    spawn,
+    SpawnOptionsWithoutStdio,
+} from "child_process";
 import * as https from "https";
 import * as zlib from "zlib";
-import { Terminal } from "./index";
+import * as unzip from "unzip-stream";
+import request from "request";
+import {Terminal} from "./index";
 
 export function executableName(name: string): string {
     return `${name}${os.platform() === "win32" ? ".exe" : ""}`;
@@ -15,16 +19,71 @@ export function changeExt(path: string, newExt: string): string {
     return path.replace(/\.[^/.]+$/, newExt);
 }
 
+export async function loadBinaryVersions(name: string): Promise<string[]> {
+    const info = await httpsGetJson(`https://binaries.tonlabs.io/${name}.json`);
+    const versions = info[name].sort(compareVersions).reverse();
+    return versions.length < 10 ? versions : [...versions.slice(0, 10), "..."];
+}
+
+async function installGlobally(dstPath: string, version: string, terminal: Terminal): Promise<void> {
+    const binDir = path.dirname(dstPath);
+    const [name, ext] = path.basename(dstPath).split(".");
+    try {
+        fs.writeFileSync(
+            `${binDir}/package.json`,
+            JSON.stringify(
+                {
+                    name: name, // ex: tonos-cli
+                    version,
+                    bin: `./${name}${ext ? "." + ext : ""}`,
+                },
+                null,
+                2,
+            ),
+        );
+        await run("npm", ["install", "-g"], {cwd: binDir}, terminal);
+    } catch (err) {
+        terminal.writeError(err);
+        throw Error(`An error occurred while trying to install ${name} globally.
+Make sure you can execute 'npm i <package> -g' without using sudo and try again`,
+        );
+    }
+}
+
+function downloadAndUnzip(dst: string, url: string): Promise<void> {
+    return new Promise((resolve, reject) => {
+        request(url)
+        .on("error", reject) // http protocol errors
+        .pipe(
+            unzip
+            .Extract({path: dst})
+            .on("error", reject) // unzip errors
+            .on("close", resolve),
+        );
+    });
+}
+
+export async function downloadFromGithub(terminal: Terminal, srcUrl: string, dstPath: string) {
+    terminal.write(`Downloading from ${srcUrl} to ${dstPath} ...`);
+    if (!fs.existsSync(dstPath)) {
+        fs.mkdirSync(dstPath, {recursive: true});
+    }
+    await downloadAndUnzip(dstPath, srcUrl);
+    terminal.write("\n");
+}
+
 function downloadAndGunzip(dest: string, url: string): Promise<void> {
     return new Promise((resolve, reject) => {
-        const request = http.get(url, response => {
+        const request = https.get(url, response => {
             if (response.statusCode !== 200) {
-                reject({
-                    message: `Download from ${url} failed with ${response.statusCode}: ${response.statusMessage}`,
-                });
+                reject(
+                    new Error(
+                        `Download from ${url} failed with ${response.statusCode}: ${response.statusMessage}`,
+                    ),
+                );
                 return;
             }
-            let file: fs.WriteStream | null = fs.createWriteStream(dest, { flags: "w" });
+            let file: fs.WriteStream | null = fs.createWriteStream(dest, {flags: "w"});
             let opened = false;
             const failed = (err: Error) => {
                 if (file) {
@@ -40,9 +99,7 @@ function downloadAndGunzip(dest: string, url: string): Promise<void> {
             const unzip = zlib.createGunzip();
             unzip.pipe(file);
 
-
             response.pipe(unzip);
-
 
             request.on("error", err => {
                 failed(err);
@@ -68,30 +125,55 @@ function downloadAndGunzip(dest: string, url: string): Promise<void> {
             });
         });
     });
-
 }
 
 export async function downloadFromBinaries(
     terminal: Terminal,
     dstPath: string,
     src: string,
-    options?: { executable?: boolean },
+    options?: {
+        executable?: boolean,
+        globally?: boolean,
+        version?: string,
+    },
 ) {
     src = src.replace("{p}", os.platform());
-    const srcUrl = `http://sdkbinaries.tonlabs.io/${src}.gz`;
+    const srcExt = path.extname(src).toLowerCase();
+    const srcUrl = `https://binaries.tonlabs.io/${src}`;
     terminal.write(`Downloading from ${srcUrl} to ${dstPath} ...`);
     const dstDir = path.dirname(dstPath);
     if (!fs.existsSync(dstDir)) {
-        fs.mkdirSync(dstDir, { recursive: true });
+        fs.mkdirSync(dstDir, {recursive: true});
     }
-    await downloadAndGunzip(dstPath, srcUrl);
+    if (srcExt === ".zip") {
+        await downloadAndUnzip(dstDir, srcUrl);
+    } else if (srcExt === ".gz") {
+        await downloadAndGunzip(dstPath, srcUrl);
+    } else {
+        throw Error(`Unexpected binary file extension: ${srcExt}`);
+    }
     if (options?.executable && os.platform() !== "win32") {
         fs.chmodSync(dstPath, 0o755);
+    }
+    if (options?.globally) {
+        if (!options.version) {
+            throw Error("Version required to install package");
+        }
+        await installGlobally(dstPath, options.version, terminal).catch(err => {
+            fs.unlink(dstPath, () => {
+            });
+            throw err;
+        });
     }
     terminal.write("\n");
 }
 
-export function run(name: string, args: string[], options: SpawnOptionsWithoutStdio, terminal: Terminal): Promise<string> {
+export function run(
+    name: string,
+    args: string[],
+    options: SpawnOptionsWithoutStdio,
+    terminal: Terminal,
+): Promise<string> {
     return new Promise((resolve, reject) => {
         try {
             const isWindows = os.platform() === "win32";
@@ -104,7 +186,6 @@ export function run(name: string, args: string[], options: SpawnOptionsWithoutSt
                     env: process.env,
                     ...options,
                 });
-            const errors: string[] = [];
             const output: string[] = [];
 
             spawned.stdout.on("data", function (data) {
@@ -113,21 +194,20 @@ export function run(name: string, args: string[], options: SpawnOptionsWithoutSt
                 terminal.log(text);
             });
 
-            spawned.stderr.on("data", (data) => {
+            spawned.stderr.on("data", data => {
                 const text = data.toString();
-                errors.push(text);
                 terminal.writeError(text);
             });
 
-            spawned.on("error", (err) => {
+            spawned.on("error", err => {
                 reject(err);
             });
 
-            spawned.on("close", (code) => {
+            spawned.on("close", code => {
                 if (code === 0) {
                     resolve(output.join(""));
                 } else {
-                    reject(errors.join(""));
+                    reject(`${name} failed`);
                 }
             });
         } catch (error) {
@@ -139,7 +219,10 @@ export function run(name: string, args: string[], options: SpawnOptionsWithoutSt
 export function uniqueFilePath(folderPath: string, namePattern: string): string {
     let index = 0;
     while (true) {
-        const filePath = path.resolve(folderPath, namePattern.replace("{}", index === 0 ? "" : index.toString()));
+        const filePath = path.resolve(
+            folderPath,
+            namePattern.replace("{}", index === 0 ? "" : index.toString()),
+        );
         if (!fs.existsSync(filePath)) {
             return filePath;
         }
@@ -168,40 +251,33 @@ export const nullTerminal: Terminal = {
     },
 };
 
-export function stringTerminal(): Terminal & { output: string, error: string } {
-    return {
-        output: "",
-        error: "",
-        write(text: string) {
-            this.output += text;
-        },
-        writeError(text: string) {
-            this.error += text;
-        },
-        log(...args: any[]) {
-            this.write(`${args.map(x => `${x}`).join(" ")}\n`);
-        },
-    };
-}
-
 export function versionToNumber(s: string): number {
     if (s.toLowerCase() === "latest") {
         return 1_000_000_000;
     }
-    const parts = `${s || ''}`.split('.').map(x => Number.parseInt(x)).slice(0, 3);
+    const parts = `${s || ""}`
+    .split(".")
+    .map(x => Number.parseInt(x))
+    .slice(0, 3);
     while (parts.length < 3) {
         parts.push(0);
     }
     return parts[0] * 1000000 + parts[1] * 1000 + parts[2];
 }
 
-let _progressLine: string = '';
+export function compareVersions(a: string, b: string): number {
+    const an = versionToNumber(a);
+    const bn = versionToNumber(b);
+    return an < bn ? -1 : (an === bn ? 0 : 1);
+}
+
+let _progressLine: string = "";
 
 export function progressLine(terminal: Terminal, line: string) {
     terminal.write(`\r${line}`);
     const extra = _progressLine.length - line.length;
     if (extra > 0) {
-        terminal.write(' '.repeat(extra) + '\b'.repeat(extra));
+        terminal.write(" ".repeat(extra) + "\b".repeat(extra));
     }
     _progressLine = line;
 }
@@ -211,24 +287,24 @@ export function progress(terminal: Terminal, info: string) {
 }
 
 export function progressDone(terminal: Terminal) {
-    terminal.log(' ✓');
-    _progressLine = '';
+    terminal.log(" ✓");
+    _progressLine = "";
 }
-
 
 export function httpsGetJson(url: string): Promise<any> {
     return new Promise((resolve, reject) => {
         const tryUrl = (url: string) => {
-            https.get(url, function (res) {
-                let body = '';
+            https
+            .get(url, function (res) {
+                let body = "";
 
-                res.on('data', function (chunk) {
+                res.on("data", function (chunk) {
                     body += chunk;
                 });
 
-                res.on('end', function () {
+                res.on("end", function () {
                     if (res.statusCode === 301) {
-                        const redirectUrl = res.headers['location'];
+                        const redirectUrl = res.headers["location"];
                         if (redirectUrl) {
                             tryUrl(redirectUrl);
                         } else {
@@ -239,12 +315,13 @@ export function httpsGetJson(url: string): Promise<any> {
                     const response = JSON.parse(body);
                     resolve(response);
                 });
-            }).on('error', (error) => {
+            })
+            .on("error", error => {
                 reject(error);
             });
         };
         tryUrl(url);
-    })
+    });
 }
 
 function toIdentifier(s: string): string {